refactor: Rename `lib/isAuthenticated.js`
