+++ lib/dashboard.js
@@ -2,7 +2,7 @@
 
 const open = require('open');
 const { getPlatformClientWithAccessKey } = require('./client-utils');
-const isAuthenticated = require('./isAuthenticated');
+const isAuthenticated = require('./is-authenticated');
 const log = require('./log');
 
 const dashboardUrl =
