+++ lib/dashboard.test.js
@@ -24,7 +24,7 @@ const { dashboardHandler, getDashboardProvidersUrl } = proxyquire('./dashboard',
   './client-utils': {
     getPlatformClientWithAccessKey: async () => new ServerlessSDKMock(),
   },
-  './isAuthenticated': () => isAuthenticated,
+  './is-authenticated': () => isAuthenticated,
   'open': openStub,
 });
