+++ lib/is-authenticated.js
