+++ lib/outputCommand.js
@@ -3,7 +3,7 @@
 const { isEmpty, isObject } = require('lodash');
 const { writeText } = require('@serverless/utils/log');
 const log = require('./log');
-const isAuthenticated = require('./isAuthenticated');
+const isAuthenticated = require('./is-authenticated');
 const throwAuthError = require('./throwAuthError');
 const resolveOutputs = require('./resolveOutputs');
 const resolveOutput = require('./resolveOutput');
