+++ lib/plugin.js
@@ -22,7 +22,7 @@ const { getDashboardUrl, dashboardHandler } = require('./dashboard');
 const setApiGatewayAccessLogFormat = require('./setApiGatewayAccessLogFormat');
 const paramCommand = require('./paramCommand');
 const outputCommand = require('./outputCommand');
-const isAuthenticated = require('./isAuthenticated');
+const isAuthenticated = require('./is-authenticated');
 const throwAuthError = require('./throwAuthError');
 
 const unconditionalCommands = new Set([
