+++ lib/plugin.test.js
@@ -51,7 +51,7 @@ describe('plugin', () => {
       './generate-event': { eventDict: {}, generate },
       './deploy-profile': { configureDeployProfile },
       './inject-logs-iam-role': injectLogsIamRole,
-      './isAuthenticated': () => true,
+      './is-authenticated': () => true,
       './setApiGatewayAccessLogFormat': setApiGatewayAccessLogFormat,
       '@serverless/config/utils': {
         getLoggedInUser: () => ({
