+++ lib/resolveParams.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const memoizee = require('memoizee');
 const { serviceSlug, instanceSlug } = require('./utils');
 const { getPlatformClientWithAccessKey } = require('./client-utils');
-const isAuthenticated = require('./isAuthenticated');
+const isAuthenticated = require('./is-authenticated');
 
 const mapDashboardParamType = (paramType) => {
   // "paramType" may not be guaranteed for old deprecated params stored on deployment profile
